move useAvailableSegments to related comp
