+++ src/ducks/Watchlists/Widgets/Filter/Metric/MarketSegments/index.js
@@ -8,7 +8,7 @@ import Icon from '@santiment-network/ui/Icon'
 import Skeleton from '../../../../../../components/Skeleton/Skeleton'
 import ExplanationTooltip from '../../../../../../components/ExplanationTooltip/ExplanationTooltip'
 import MetricState from '../MetricState'
-import { useAvailableSegments } from '../../../../gql/hooks'
+import { useAvailableSegments } from './hooks'
 import { extractFilterByMetricType } from '../../detector'
 import Suggestions from '../Suggestions'
 import { filterSegmentsBySearch } from './utils'
